+++ java/cleanarchitecture/src/test/java/domain/PhoneNumberTest.java
@@ -1,23 +1,23 @@
-package school;
+package domain;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
-import com.cleanarchitecture.school.Email;
+import com.cleanarchitecture.domain.school.PhoneNumber;
 
-public class EmailTest {
+public class PhoneNumberTest {
 
     @Test
     void InvalidArguments() {
         assertThrows(IllegalArgumentException.class, 
-            () -> new Email(null));
+            () -> new PhoneNumber(null, null));
 
         assertThrows(IllegalArgumentException.class, 
-            () -> new Email(""));
+            () -> new PhoneNumber("", ""));
 
         assertThrows(IllegalArgumentException.class, 
-            () -> new Email("teste"));
+            () -> new PhoneNumber("24", "91111"));
             
     }
 }
